replace manual code with generated code for p_XLShiftZ_URShiftZ.

+++ src/hotspot/share/opto/mulnode.cpp
@@ -1276,7 +1276,7 @@ if (_JOG_in1->Opcode() == Op_LShiftI
     && _JOG_in2->Opcode() == Op_ConI
     && _JOG_in12 == _JOG_in2) {
 jint z = phase->type(_JOG_in2)->isa_int()->get_con();
-return new AndINode(_JOG_in11, phase->intcon(java_subtract(1 << java_subtract(32, z), 1)));
+return new AndINode(_JOG_in11, phase->intcon(java_subtract(1 << java_subtract((jint) 32, z), (jint) 1)));
 }
 }
